feat: add Jotai global state management guide and reorganize theme-related files
