+++ src/hooks/index.ts
@@ -1,0 +1,2 @@
+// Barrel exports for all hooks
+export * from './useTheme'
